+++ v3/changelog.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Added configuration for headers in response to client for requests handled by
+  the pre-route plugin hook.
+
 ### Fixed
 
 - Fixed a bug where TypePermissions field presets would be ignored if the type